Clean up tree_dresser.rb
- Move SexpExtensions and SexpNode into their own files
- Cleaner extension lookup
- Fix code smells
